rename: bid-form 폴더 생성 및 컴포넌트 정리
