+++ apps/web/src/components/auction-detail/bid-form/bottom.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const BidFormBottom = () => {
+  return <div>BidFormBottom</div>;
+};
+
+export default BidFormBottom;
