+++ apps/web/src/components/auction-detail/index.tsx
@@ -1,6 +1,3 @@
-export { default as BannerDirect } from './banner-direct';
-export { default as BidForm } from './bid-form';
-export { default as BidInput } from './bid-input';
 export { default as ButtonDetailMore } from './button-detail-more';
 export { default as ButtonFavorite } from './button-favorite';
 export { default as CountDown } from './countdown';
@@ -14,4 +11,8 @@ export { default as BidTableHead } from './bid-table/head';
 export { default as BidTableRow } from './bid-table/row';
 export { default as BidderAvatar } from './bid-table/bidder-avatar';
 
-// export { default as BidTable } from './bid-table';
+//bid-form
+export { default as BidForm } from './bid-form/form';
+export { default as BidFormInput } from './bid-form/input';
+export { default as BidFormBottom } from './bid-form/bottom';
+export { default as ButtonDirectDeal } from './bid-form/button-direct-deal';
